Build PHPStan image after its parent completes
The automated build for the PHPStan image failed because it depends on the
integration test image, which didn't exist at the time. Moments later, it did
exist, so this build will work next time it runs.

This change allows the build to work properly, and also rebuilds these images
when the upstream/parent image gets any changes - like adding a PHP extension.